Added MySQL connection, working login screen (SQLi safe via prepared statements), models for patient and staff, fixed graph to show date instead just a number, and also only shows the 10 most recent data from each series.

+++ src/main/java/com/hmaar/sundhed/model/implementation/EKG.java
@@ -4,9 +4,9 @@ import com.hmaar.sundhed.model.interfaces.EKGData;
 
 public class EKG implements EKGData {
     private double voltage;
-    private double time;
+    private long time;
 
-    public EKG(double voltage, double time) {
+    public EKG(double voltage, long time) {
         this.time=time;
         this.voltage=voltage;
 
@@ -23,12 +23,12 @@ public class EKG implements EKGData {
     }
 
     @Override
-    public double getTime() {
+    public long getTime() {
         return time;
     }
 
     @Override
-    public void setTime(double time) {
+    public void setTime(long time) {
         this.time = time;
     }
 }
