+++ src/main/java/com/hmaar/sundhed/model/interfaces/PulsData.java
@@ -3,6 +3,6 @@ package com.hmaar.sundhed.model.interfaces;
 public interface PulsData {
     void setPuls(double puls);
     double getPuls();
-    void setTime(double time);
-    double getTime();
+    void setTime(long time);
+    long getTime();
 }
