+++ src/main/java/com/hmaar/sundhed/model/interfaces/SpO2Data.java
@@ -3,7 +3,7 @@ package com.hmaar.sundhed.model.interfaces;
 public interface SpO2Data {
     void setSpO2(double spO2);
     double getSpO2();
-    void setTime(double time);
-    double getTime();
+    void setTime(long time);
+    long getTime();
 }
