+++ src/main/java/com/hmaar/sundhed/model/interfaces/TempData.java
@@ -3,6 +3,6 @@ package com.hmaar.sundhed.model.interfaces;
 public interface TempData {
     void setTemp(double voltage);
     double getTemp();
-    void setTime(double time);
-    double getTime();
+    void setTime(long time);
+    long getTime();
 }
